feat(appointments): refactored appointment detail form to own component
